- Added the types of responses for controllers - Fixed the proccesses of some methods in policy and dependency services - Quick check on all system files to remove unnecessary lines, like unused directive and so on.

+++ Backend/Makassed.Api/Services/Policies/PolicyService.cs
@@ -37,21 +37,37 @@ public class PolicyService : IPolicyService
         return policy is null;
     }
 
-    public async Task<List<Policy>> GetPoliciesAsync(SieveModel sieveModel, Guid chapterId)
+    private async Task<bool> CheckChapterExists(Guid id) 
+    { 
+        var chapter = await _chapterRepository.GetChapterByIdAsync(id);
+
+        return chapter is not null;
+    }
+
+    // check if policy belongs to the chapter
+    private async Task<bool> CheckPolicyBelongsToChapter(Guid chapterId, Guid policyId)
     {
+        var policy = await _policyRepository.GetPolicyByIdAsync(policyId);
+
+        return policy?.ChapterId == chapterId;
+    }
+
+    public async Task<ErrorOr<List<Policy>>> GetPoliciesAsync(SieveModel sieveModel, Guid chapterId)
+    {
+        if (!await CheckChapterExists(chapterId))
+            return Errors.Chapter.NotFound;
+
         return await _policyRepository.GetPoliciesAsync(sieveModel, chapterId);
     }
 
-    public async Task<ErrorOr<Policy>> GetPolicyByIdAsync(Guid id)
+    public async Task<ErrorOr<Policy>> GetPolicyByIdAsync(Guid chapterId, Guid id)
     {
+        if (!await CheckChapterExists(chapterId))
+            return Errors.Chapter.NotFound;
+
         var policy = await _policyRepository.GetPolicyByIdAsync(id);
         
         return policy is null ? Errors.Policy.NotFound : policy;
-    }
-
-    private async Task<Chapter?> CheckChapterExists(Guid id) 
-    { 
-        return await _chapterRepository.GetChapterByIdAsync(id);
     }
 
     public async Task<ErrorOr<Created>> CreatePolicyAsync(Guid chapterId, Policy policy)
@@ -61,9 +77,9 @@ public class PolicyService : IPolicyService
         if (userRole == null)
             return Errors.User.Unauthorized;
 
-        var existedChapterResult = await CheckChapterExists(chapterId);
+        var existedChapter = await _chapterRepository.GetChapterByIdAsync(chapterId);
 
-        if (existedChapterResult is null)
+        if (existedChapter is null)
             return Errors.Chapter.NotFound;
 
         if (!await IsUniqueName(policy.Name))
@@ -82,15 +98,18 @@ public class PolicyService : IPolicyService
         
         await _policyRepository.CreatePolicyAsync(policy);
 
-        await _chapterRepository.UpdateChapterEnableStateAsync(existedChapterResult.Id);
+        await _chapterRepository.UpdateChapterEnableStateAsync(existedChapter.Id);
 
         await _unitOfWork.SaveChangesAsync();
 
         return Result.Created;
     }
 
-    public async Task<ErrorOr<Deleted>> DeletePolicyAsync(Guid id)
+    public async Task<ErrorOr<Deleted>> DeletePolicyAsync(Guid chapterId, Guid id)
     {
+        if (!await CheckChapterExists(chapterId))
+            return Errors.Chapter.NotFound;
+
         var deletedPolicy = await _policyRepository.DeletePolicyAsync(id);
 
         if (deletedPolicy is null)
@@ -101,8 +120,11 @@ public class PolicyService : IPolicyService
         return Result.Deleted;
     }
 
-    public async Task<ErrorOr<Updated>> UpdatePolicyAsync(Guid id, Policy policy)
+    public async Task<ErrorOr<Updated>> UpdatePolicyAsync(Guid chapterId, Guid id, Policy policy)
     {
+        if (await CheckPolicyBelongsToChapter(chapterId, id))
+            return Errors.Policy.DoesNotBelongToChapter;
+
         policy.PdfUrl = await _localFileStorageService.UploadFileAndGetUrlAsync(policy.MainFile);
         policy.PageCount = _localFileStorageService.GetPdfFilePageCount(policy.MainFile);
